+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - let/context DRA for filter/order
 - hide/show filters
 - results.name nesecarry? ?? : for spec func
+- spec: Category before accept all
 
 ### new feature
 - embed youtube in score (or just link to search page ?)